Merge in EDGEAI-ALGO/edgeai-mmdetection3d from develop to main

* commit '1a63408e86239ef7dcb398876d9db6321bbfa926':
  [BEV Doc] Updated BEV Documents (README's)
  [PandaSet] EDGEAI_ALGO-205: 1. Added PandaSet configs and dataset for PETRv2, Far3D and StreamPETR 2. Updated pandaset_converter.py for StreamPETR including bug fixes 3. Fixed errors in BEVDet and PETR that were introduced by code    structuring and Sparse4D changes.

+++ projects_edgeai/PETR/petr/onnx_export.py
@@ -37,7 +37,7 @@ def export_PETR(model, inputs=None, data_samples=None,
             is_prev_feat = 0
         else:
             is_prev_feat = 1
-        is_prev_feat = torch.tensor(is_prev_feat, dtype=torch.float32, device=img.device) 
+        is_prev_feat = torch.tensor(is_prev_feat, dtype=torch.int32, device=img.device)
 
     batch_img_metas = onnxModel.add_lidar2img(img, batch_img_metas)
     onnxModel.prepare_data(img, batch_img_metas)
